# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/notification-recipients:
    get:
      operationId: adminListNotificationRecipients
      tags: [auth]
      summary: List all notification recipient assignments (super admin only)
      responses:
        "200":
          description: List of notification recipient assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationRecipient"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateNotificationRecipient
      tags: [auth]
      summary: Assign an admin user to a notification module (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationRecipientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationRecipient"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /enquiries:
    post:
      operationId: createEnquiryPublic
      tags: [engagement]
      summary: Submit a new enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertEnquiry"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/enquiries:
    get:
      operationId: adminListEnquiries
      tags: [engagement]
      summary: List all enquiries (admin)
      responses:
        "200":
          description: List of enquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enquiry"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/volunteer-opportunities:
    get:
      operationId: adminListVolunteerOpportunities
      tags: [engagement]
      summary: List all volunteer opportunities (admin)
      responses:
        "200":
          description: List of volunteer opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VolunteerOpportunity"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateVolunteerOpportunity
      tags: [engagement]
      summary: Create a new volunteer opportunity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertVolunteerOpportunity"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerOpportunity"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /volunteer-applications:
    post:
      operationId: createVolunteerApplicationPublic
      tags: [engagement]
      summary: Submit a new volunteer application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertVolunteerApplication"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerApplication"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/volunteer-applications:
    get:
      operationId: adminListVolunteerApplications
      tags: [engagement]
      summary: List all volunteer applications (admin)
      responses:
        "200":
          description: List of volunteer applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VolunteerApplication"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    NotificationModule:
      type: string
      enum:
        - donations
        - enquiries
        - courses
        - volunteers
    NotificationRecipient:
      type: object
      properties:
        id:
          type: string
        adminUserId:
          type: string
        module:
          $ref: "#/components/schemas/NotificationModule"
        emailEnabled:
          type: boolean
        smsEnabled:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - adminUserId
        - module
        - emailEnabled
        - smsEnabled
        - createdAt
        - updatedAt
    CreateNotificationRecipientInput:
      type: object
      properties:
        adminUserId:
          type: string
        module:
          $ref: "#/components/schemas/NotificationModule"
        emailEnabled:
          type: boolean
        smsEnabled:
          type: boolean
      required:
        - adminUserId
        - module
...
    Enquiry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - subject
        - message
        - status
        - createdAt
    InsertEnquiry:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        status:
          type: string
      required:
        - name
        - email
        - subject
        - message
...
    VolunteerOpportunity:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - active
        - createdAt
    InsertVolunteerOpportunity:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
      required:
        - title
        - description
        - active
...
    VolunteerApplication:
      type: object
      properties:
        id:
          type: string
        opportunityId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - opportunityId
        - name
        - email
        - status
        - createdAt
    InsertVolunteerApplication:
      type: object
      properties:
        opportunityId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
      required:
        - opportunityId
        - name
        - email
...