# lib/api-zod/src/generated/types/createNotificationRecipientInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationModule } from './notificationModule';

export interface CreateNotificationRecipientInput {
  adminUserId: string;
  module: NotificationModule;
  emailEnabled?: boolean;
  smsEnabled?: boolean;
}