# lib/api-zod/src/generated/types/notificationModule.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type NotificationModule = typeof NotificationModule[keyof typeof NotificationModule];


export const NotificationModule = {
  donations: 'donations',
  enquiries: 'enquiries',
  courses: 'courses',
  volunteers: 'volunteers',
} as const;